// sp/ClientApp/src/app/models/members/profile-employment.model.ts
export class MemberProfileEmploymentModel {
    EmpInfoID: string;
    City: string;
    State: string;
    StartMonth: string;
    StartYear: string;
    EndMonth: string;
    EndYear: string;
    CurrentlyWorkedHere: string;
    companyID: string;
    companyName: string;
    companyImage: string;
    companyAddress: string;
    title: string;
    datesWorked: string;
    Description: string;
    Email: string;
    Industry: string;
    IPOYear: string;
    JobDesc: string;
    Sector: string;
    summaryQuote: string;
    Symbol: string;
    website: string;
}    